requirements.txt
python-telegram-bot==21.10
yt-dlp==2025.3.31
python-dotenv==1.1.0
